Fix clean_extraction deal with Wf

+++ safechecker-plugin/clean_extraction.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+SED=`which gsed | which sed`
+
 echo "Cleaning result of extraction"
 
 if [ ! -d "src" ]
@@ -30,9 +32,9 @@ then
     cd ..
 
     # confusion between Init.Wf and Program.Wf
-    mv src/wf.ml src/wf0.ml
-    mv src/wf.mli src/wf0.mli
-    sed -i.bak src/pCUICSafeChecker.ml -e 's/open Wf/open Wf0/'
+    if [ -f src/wf.ml ]; then mv src/wf.ml src/wf0.ml; fi;
+    if [ -f src/wf.mli ]; then mv src/wf.mli src/wf0.mli; fi;
+    ${SED} -i -e "s/open Wf/open Wf0/" src/pCUICSafeChecker.ml
 
     # Remove extracted modules already linked in the template_coq plugin.
     echo "Removing:" $files
